+++ RaceControl.slnx
@@ -2,5 +2,6 @@
   <Folder Name="/Tests/">
     <Project Path="tests/RaceControl.Tests/RaceControl.Tests.csproj" />
   </Folder>
+  <Project Path="src/RaceControl.Database/RaceControl.Database.csproj" />
   <Project Path="src/RaceControl/RaceControl.csproj" />
 </Solution>
