--- app.json
@@ -1,14 +1,0 @@
-{
-  "healthchecks": {
-    "web": [
-      {
-        "type": "liveness",
-        "name": "web check",
-        "description": "Checking if the app is responding to /health",
-        "path": "/health",
-        "port": 5000,
-        "attempts": 3
-      }
-    ]
-  }
-}
