+++ src/RaceControl.Database/Entities/Category.cs
