+++ src/RaceControl.Database/Entities/Session.cs
